[refactor] ActionTest로 통합
